FixtureFileWatcher working
